refactor: creating folders and refactoring service files
